+++ example/active-learning.py
@@ -3,7 +3,11 @@
 import numpy as np
 import matplotlib.pyplot as plt
 from graphdot.model.gaussian_process import GaussianProcessRegressor
-from graphdot.model.active_learning import StochasticVolumeMaximizer
+from graphdot.model.active_learning import (
+    DeterminantMaximizer,
+    VarianceMinimizer,
+    HierarchicalDrafter
+)
 
 
 # Good old Gaussian/RBF kernel
@@ -30,8 +34,11 @@ kernel = Kernel(s=2.0)
 X = np.sort(np.random.randn(300) * 10)
 y = f(X)
 n = 40  # training set budget
-vm = StochasticVolumeMaximizer(kernel)
-active_set = vm(X, n)
+if False:
+    drafter = HierarchicalDrafter(DeterminantMaximizer(kernel))
+else:
+    drafter = HierarchicalDrafter(VarianceMinimizer(kernel))
+active_set = drafter(X, n)
 random_set = np.random.choice(len(X), n, False)
 
 
@@ -45,7 +52,8 @@ def test(chosen, label, color):
 
 plt.figure()
 grid = np.linspace(X.min(), X.max(), 500)
-plt.plot(grid, f(grid), lw=0.5, color='k', label='ground_truth')
+plt.plot(grid, f(grid), lw=2, ls='dashed', color='k', alpha=0.5,
+         label='ground_truth')
 test(active_set, 'active', (0.0, 0.2, 0.7))
 test(random_set, 'random', (0.9, 0.4, 0.0))
 plt.legend()